backend/requirements.txt
# Deep Thinking Backend - Requirements
# A lightweight LLM service for conversational AI with multiple thinking depths

# Core Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0

# LLM Providers
openai>=1.60.0
anthropic>=0.40.0
httpx>=0.28.0

# Utilities
python-dotenv>=1.0.1
pyyaml>=6.0.2

# Optional: Redis caching (uncomment if needed)
# redis>=5.0.0

# Optional: Prometheus metrics (uncomment if needed)
# prometheus-client>=0.21.0
